Fix an issue with loading data from page to page
If the Guides were already loaded into the guide collection, the
sync callback wasn't getting called on the new page load. This
caused the data to never show up.

This change just sets the data based on existing data in the Guides
Collection if there is any.

Also, I moved more logic into the collection and renamed the methods
to represent that they return Guide objects instead of just plan
objects.