Move category definition to the middleware
Categories don't change often and we use them throughout the site. It
seemed unnecessary to have to make them a passed dependency for each
react component instead of just being able to reference the information
as necessary from an already defined instance.

This moves the categories instance into a single instance and uses the
middleware layer to make sure that it's loaded before the site loads