Add the user photo and visual state to the header nav
Logout is pretty straight forward, but login is tricky because we
offer facebook and twitter login. Until we have a prompt, it doesn't
make much sense to have a login button there. (especially with the
auth bar that comes down from the top).

For the time being, I've removed the button for when a user isn't
logged in.

Per #62