Rough first pass at page titles
For now, I've just set a standard page title and used the name of
the page at the end. I think a better solution would be to have
custom page titles based on the route or the guide properties.

Hoever, there is a bunch of work that should be done for SEO support
and I think that could be done during the cycle.

Per #128