calling the GuideModules export as a function call
our understanding of this is that a require cache kicks in to avoid require cycles. This causes GuideModules to return partially formed objects and not the expected module array.

According to documentation, it’s best to export it as a function, so that the cache doesn’t come in the way.
http://nodejs.org/docs/latest/api/modules.html#modules_caching