Show the current user count on the splash screen
This is a pretty ineffecient query right now. Because of Firebase's
inability to do the count on the server, we're forced to load all
of the users and then count them on the client side. UGLY.

To work around this, we can store a counter on firebase of the current
number of users. I'll create a new issue for this to keep track of it.

Per #119