Show login prompt after user returns to site
This actually shows the login prompt when the user clicks on any link
that has a target of "_blank", it just happens that the user won't
see it until they come back to the site.

The autolinker that we use automatically adds target="_blank" to
the links that it creates, so for all standard text links, this is
covered.

For all other links, we need to manually add target="_blank" to them.

Closes #196