Redirect user on first visit
If the user has been to the site before and lands on the home page,
redirect them to the guides grid immediately.

If the user has not been to the site and lands on the /guides page,
redirect them to the splash page immediately.

We're using localStorage to track the firstVisit because it's less
request intensive than cookies and has all the browser support we
need.

Closes #222