Redirect Fort Collins users to the city page
If the user hasn't visited before, and they land on the homepage
we check if they are from Fort Collins and if so we redirect them
to the fort collins page.

This does not work on the server side, so we need to figure out what
that experience looks like. We may have to do a server side redirect.

Per #302